tennisweb_backend/requirements.txt
django==5.2.6
# Use a stable DRF version available on PyPI that is compatible
djangorestframework==3.16.1
django-cors-headers==4.0.0
djangorestframework-simplejwt==5.5.1
Pillow==11.0.0
